test(reconciliation): restore migration CI coverage
