+++ scripts/ci_backend_test_manifest.json
@@ -573,6 +573,7 @@
         "tests::upstream_reconciliation_key_observations::",
         "tests::upstream_reconciliation_fencing::",
         "tests::upstream_reconciliation_projection::",
+        "tests::upstream_reconciliation_source_identity::",
         "tests::upstream_reconciliation::reconciliation_controlled_retry_advances_the_claim_attempt_before_success",
         "tavily_proxy::reconciliation_engine_tests::",
         "tavily_proxy::user_business_calls_memory::memory_window_regression_tests::",
