+++ src/tests/mod.rs
@@ -52,6 +52,7 @@ mod upstream_reconciliation_fencing;
 mod upstream_reconciliation_key_observations;
 mod upstream_reconciliation_observation;
 mod upstream_reconciliation_projection;
+mod upstream_reconciliation_source_identity;
 mod usage_series_and_backfills;
 mod user_business_calls_1h;
 mod user_pressure_analysis;
